+++ CHANGELOG.md
@@ -36,7 +36,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 - Drop the feature for abbreviated date ranges (#1057)
 - Drop time zones and use floating dates for iCal (#1054)
-- Remove the file upload from the FE editor (#1053)
+- Remove the file upload from the FE editor (#1053, #1058)
 - Drop the unused `TimeRangeViewHelper` (#1006)
 - Drop the fax number field from the speakers (#999)
 - Drop autogeneratable fields from the SQL file (#905)